pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mg"
version = "3.0.3"
authors = [
    {name = "Gabe Garrido", email = "gabriel.garrido28@gmail.com"},
]
description = "Master Global package."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"  # Use the SPDX identifier
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas>=2.0.0",
    "tabulate==0.8.10",
    "requests==2.29.0",
    "google-api-python-client==2.88.0",
    "sqlalchemy==1.4.47",
    "python-tds==1.12.0",
    "sqlalchemy-pytds==0.3.5",
    "pyodbc==4.0.39",
    "postgres==4.0",
    "psycopg2==2.9.7",
    "unidecode==1.3.8",
    "psutil==6.0.0",
    "python-Levenshtein==0.26.0",
    "marshmallow==3.23.1",
    "GitPython==3.1.44",
    "gcloud-aio-auth==5.3.2",
    "gcloud-aio-storage==9.3.0",
    "dill==0.3.9",
    "google-cloud-run==0.10.16",
    "google-cloud-secret-manager==2.17.0",
    "google-cloud-pubsub==2.29.0"
    ]

# Correct format is a direct array, not a nested structure
license-files = ["LICENSE"]